Issue #11
The web root often contains temporary or working directories that don't
need to be included in the stale sites report. This commit added
an "ignored sites" feature that will allow someone viewing the
stale sites report to quickly exclude sites from the report
There is a list of ignored sites displayed, and users are able to un-ignore
a site if they ignored it accidentally. The list of ignored directories is global,
and not per-user.

Signed-off-by:  <adarshakb@gmail.com>