Issue #22
Added a new function to check if the file name is present in the
ignored file names. The problem was that the file path was being
incorrectly matched to a file name. Changed it to compare the last
part of the path to extract the file name and the compare it.

Signed-off-by:  <adarshakb@gmail.com>